Move config watcher to config pkg
